requirements/tests.txt
pytest-asyncio
pytest-cov
pytest-env
httpx

# for mocking aiohttp
aioresponses

# for test sqlite db
aiosqlite
